pyproject.toml
[project]
name = "report-automation"
version = "0.1.0"
description = "Monthly ads reporting automation pipeline"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "google-cloud-bigquery>=3.18.0",
  "google-api-python-client>=2.116.0",
  "google-auth>=2.27.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
report-automation = "report_automation.main:main"
